Measure Weight

+++ main/ui/view01.h
@@ -3,7 +3,7 @@
 #include "lvgl.h"
 
 void make_widget_tree(lv_event_cb_t, lv_event_cb_t);
-void set_weight(int);
+void set_weight(int16_t);
 void set_timer(int);
 bool is_on();
 void toggle_on();
